fix: bsp/win32编译错误

+++ components/ui/sdl2/luat_sdl2.c
@@ -51,7 +51,7 @@ int luat_sdl2_deinit(luat_sdl2_conf_t *conf) {
     return 0;
 }
 
-void luat_sdl2_draw(size_t x1, size_t y1, size_t x2, size_t y2, uint32_t* data) {
+void luat_sdl2_draw(int16_t x1, int16_t y1, int16_t x2, int16_t y2, uint32_t* data) {
     SDL_Rect r;
     r.x = x1;
     r.y = y1;
