+++ components/ui/sdl2/luat_sdl2.h
@@ -14,7 +14,7 @@ typedef struct luat_sdl2_conf
 int luat_sdl2_init(luat_sdl2_conf_t *conf);
 int luat_sdl2_deinit(luat_sdl2_conf_t *conf);
 
-void luat_sdl2_draw(size_t x1, size_t y1, size_t x2, size_t y2, uint32_t* data);
+void luat_sdl2_draw(int16_t x1, int16_t y1, int16_t x2, int16_t y2, uint32_t* data);
 void luat_sdl2_flush(void);
 
 #endif
